+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+* Miri support: fixed an aliasing issue in `PooledBuffers` and a `transmute` flagged by miri; unit tests for covering all unsafe code; miri CI run on those unit tests
+
 ## [0.4.1] - 2026-09-17
 * Fix: `NameSliceIter::next_back` in the builtin mDNS responder always returned `None`, breaking reverse label iteration (#554)
 * Fix: BTP ACKs sender was wrongly using the RECV timeout of 15s for sending (#554)